requirements.txt
streamlit
streamlit-authenticator
PyYAML
tensorflow
numpy
pillow
matplotlib
bcrypt
pandas
